+++ structs/group.go
@@ -1,13 +1,12 @@
 package structs
 
-
 type Group struct {
-	Id int `json:"id"`
-	Name string `json:"name"`
+	Id   float64 `json:"id"`
+	Name string  `json:"name"`
 }
 
 type GroupResponseList struct {
-	Count int `json:"count"`
+	Count int     `json:"count"`
 	Items []Group `json:"items"`
 }
